+++ packages/jellyfish-api-core/__tests__/category/wallet/wallet.test.ts
@@ -1,6 +1,6 @@
-import { ContainerAdapterClient } from '../container_adapter_client'
+import { ContainerAdapterClient } from '../../container_adapter_client'
 import { MasterNodeRegTestContainer, RegTestContainer } from '@defichain/testcontainers'
-import { BigNumber, wallet } from '../../src'
+import { BigNumber, wallet } from '../../../src'
 import waitForExpect from 'wait-for-expect'
 import {
   UTXO,
@@ -9,7 +9,7 @@ import {
   SendToAddressOptions,
   Mode,
   SendManyOptions
-} from '../../src/category/wallet'
+} from '../../../src/category/wallet'
 
 describe('getBalance', () => {
   describe('regtest', () => {
@@ -874,82 +874,5 @@ describe('masternode', () => {
         expect(typeof transactionId).toStrictEqual('string')
       })
     })
-  })
-})
-
-describe('dumpPrivKey', () => {
-  const container = new MasterNodeRegTestContainer()
-  const client = new ContainerAdapterClient(container)
-
-  beforeAll(async () => {
-    await container.start()
-    await container.waitForReady()
-    await container.waitForWalletCoinbaseMaturity()
-    await container.waitForWalletBalanceGTE(101)
-  })
-
-  afterAll(async () => {
-    await container.stop()
-  })
-
-  it('should reveal private key of given address', async () => {
-    await waitForExpect(async () => {
-      const address = 'mwsZw8nF7pKxWH8eoKL9tPxTpaFkz7QeLU'
-      const privateKey = await client.wallet.dumpPrivKey(address)
-
-      expect(typeof privateKey).toStrictEqual('string')
-      expect(privateKey).toStrictEqual('cRiRQ9cHmy5evDqNDdEV8f6zfbK6epi9Fpz4CRZsmLEmkwy54dWz')
-    })
-  })
-
-  it('should throw and error when invalid DFI address is provided', async () => {
-    await waitForExpect(async () => {
-      const invalidAddress = 'invalidAddress'
-
-      await expect(client.wallet.dumpPrivKey(invalidAddress)).rejects.toThrow('Invalid Defi address')
-    })
-  })
-})
-
-describe('importPrivKey', () => {
-  const container = new MasterNodeRegTestContainer()
-  const client = new ContainerAdapterClient(container)
-
-  beforeAll(async () => {
-    await container.start()
-    await container.waitForReady()
-    await container.waitForWalletCoinbaseMaturity()
-  })
-
-  afterAll(async () => {
-    await container.stop()
-  })
-
-  it('should should import private key without failing ', async () => {
-    const privatekey = await client.wallet.dumpPrivKey(await client.wallet.getNewAddress())
-    const promise = client.wallet.importPrivKey(privatekey)
-
-    await expect(promise).resolves.not.toThrow()
-  })
-
-  it('should import private key with rescan set to false', async () => {
-    const privateKey = await client.wallet.dumpPrivKey(await client.wallet.getNewAddress())
-    const promise = client.wallet.importPrivKey(privateKey, '', false)
-
-    await expect(promise).resolves.not.toThrow()
-  })
-
-  it('should import private key with label passed', async () => {
-    const privateKey = await client.wallet.dumpPrivKey(await client.wallet.getNewAddress('testing'))
-    const promise = client.wallet.importPrivKey(privateKey, 'testing')
-
-    await expect(promise).resolves.not.toThrow()
-  })
-
-  it('should fail and throw an error with invalid private key', async () => {
-    const privateKey = 'invalidPrivateKey'
-    const promise = client.wallet.importPrivKey(privateKey)
-
-    await expect(promise).rejects.toThrow('Invalid private key encoding')
   })
 })
